Show Rich progress bars in notebooks and lib calls
