+++ kelp_o_matic/__init__.py
@@ -1,6 +1,6 @@
 from kelp_o_matic import geotiff_io, models
 from kelp_o_matic.lib import find_kelp, find_mussels
-from kelp_o_matic.managers import GeotiffSegmentationManager
+from kelp_o_matic.managers import GeotiffSegmentationManager, RichSegmentationManager
 
 __all__ = [
     "find_kelp",
@@ -8,5 +8,6 @@ __all__ = [
     "geotiff_io",
     "models",
     "GeotiffSegmentationManager",
+    "RichSegmentationManager",
 ]
 __version__ = "0.0.0"
